+++ Source/Bogus/DataSets/Lorem.cs
@@ -8,9 +8,11 @@ namespace Bogus.DataSets
    /// </summary>
    public class Lorem : DataSet
    {
+
       /// <summary>
-      /// Default constructor
+      /// Initializes a new instance of the <see cref="Lorem"/> class.
       /// </summary>
+      /// <param name="locale">The locale used to generate random values.</param>
       public Lorem(string locale = "en") : base(locale)
       {
       }
@@ -24,31 +26,32 @@ namespace Bogus.DataSets
       }
 
       /// <summary>
-      /// Get some lorem words
+      /// Get an array of random lorem words.
       /// </summary>
+      /// <param name="num">The number of random lorem words to return.</param>
       public string[] Words(int num = 3)
       {
-         return Enumerable.Range(1, num).Select(f => Word()).ToArray(); // lol
+         return Enumerable.Range(1, num).Select(_ => Word()).ToArray();
       }
 
       /// <summary>
       /// Get a character letter.
       /// </summary>
-      /// <param name="num">Number of characters to return.</param>
+      /// <param name="num">The number of characters to return.</param>
       public string Letter(int num = 1)
       {
          if( num <= 0 )
             return string.Empty;
 
-         var w = Words(1)[0];
-         var c = Random.ArrayElement(w.ToArray());
-         return c + Letter(num - 1);
+         var words = Words(1)[0];
+         var characters = Random.ArrayElement(words.ToArray());
+         return characters + Letter(num - 1);
       }
 
       /// <summary>
       /// Get a random sentence of specific number of words. 
       /// </summary>
-      /// <param name="wordCount">Get a sentence with wordCount words. Defaults between 3 and 10</param>
+      /// <param name="wordCount">Get a sentence with wordCount words. Defaults between 3 and 10.</param>
       /// <param name="range">Add anywhere between 0 to 'range' additional words to wordCount. Default is 0.</param>
       public string Sentence(int? wordCount = null, int? range = 0)
       {
@@ -65,12 +68,13 @@ namespace Bogus.DataSets
       /// <summary>
       /// Get some sentences.
       /// </summary>
-      /// <param name="sentenceCount">The number of sentences</param>
+      /// <param name="sentenceCount">The number of sentences.</param>
+      /// <param name="separator">The string to separate sentences.</param>
       public string Sentences(int? sentenceCount = null, string separator = "\n")
       {
          var sc = sentenceCount ?? this.Random.Number(2, 6);
          var sentences = Enumerable.Range(1, sc)
-            .Select(s => Sentence());
+            .Select(_ => Sentence());
 
          return string.Join(separator, sentences);
       }
@@ -89,24 +93,26 @@ namespace Bogus.DataSets
       /// <summary>
       /// Get a specified number of paragraphs.
       /// </summary>
-      /// <param name="count">Number of paragraphs</param>
+      /// <param name="count">Number of paragraphs.</param>
+      /// <param name="separator">The string to separate paragraphs.</param>
       public string Paragraphs(int count = 3, string separator = "\n\n")
       {
-         var paras = Enumerable.Range(1, count)
-            .Select(i => Paragraph());
+         var paragraphs = Enumerable.Range(1, count)
+            .Select(_ => Paragraph());
 
-         return string.Join(separator, paras);
+         return string.Join(separator, paragraphs);
       }
 
       /// <summary>
       /// Get a random number of paragraphs between <paramref name="min"/> and <paramref name="max"/>.
       /// </summary>
-      /// <param name="min">Minimum number of paragraphs</param>
-      /// <param name="max">Maximum number of paragraphs</param>
+      /// <param name="min">Minimum number of paragraphs.</param>
+      /// <param name="max">Maximum number of paragraphs.</param>
+      /// <param name="separator">The string to separate the paragraphs.</param>
       public string Paragraphs(int min, int max, string separator = "\n\n")
       {
-         var n = this.Random.Number(min, max);
-         return Paragraphs(n, separator);
+         var count = this.Random.Number(min, max);
+         return Paragraphs(count, separator);
       }
 
       /// <summary>
@@ -123,6 +129,8 @@ namespace Bogus.DataSets
       /// <summary>
       /// Get lines of lorem.
       /// </summary>
+      /// <param name="lineCount">The amount of lines to generate. Defaults between 1 and 5.</param>
+      /// <param name="separator">The string to separate the lines.</param>
       public string Lines(int? lineCount = null, string separator = "\n")
       {
          var lc = lineCount ?? this.Random.Number(1, 5);
@@ -133,6 +141,7 @@ namespace Bogus.DataSets
       /// <summary>
       /// Slugify lorem words.
       /// </summary>
+      /// <param name="wordcount">The amount of words to slugify.</param>
       public string Slug(int wordcount = 3)
       {
          var words = Words(wordcount);